changed: stuff up and what not
The code is cleaner and stuff, technically faster, and I'm aware PRE()
is a terrible implementation of what I'm trying to achieve but it's my
code and I'm not bothered. I also decided to go for POSIX compliancy
instead of Bash, and went for `[] && {}` even if it's less readable then
`if [] && then; fi`

My code so I decide.